# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refreshToken",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "name": "initialDate",
            "required": true,
            "in": "query",
            "description": "Data inicial de criação do usuario",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "finalDate",
            "required": true,
            "in": "query",
            "description": "Data final de criação do usuario",
            "schema": {
              "example": "2025-01-01",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": true,
            "in": "query",
            "schema": {
              "example": "Fulano de tal",
              "type": "string"
            }
          },
          {
            "name": "role",
            "required": true,
            "in": "query",
            "schema": {
              "example": "admin",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/farm": {
      "post": {
        "operationId": "FarmController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFarmDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Farm"
        ]
      },
      "get": {
        "operationId": "FarmController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number to be returned (starts at 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (maximum 100)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaginationBodyDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Farm"
        ]
      }
    },
    "/farm/{id}": {
      "get": {
        "operationId": "FarmController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Farm"
        ]
      },
      "patch": {
        "operationId": "FarmController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFarmDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Farm"
        ]
      },
      "delete": {
        "operationId": "FarmController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Farm"
        ]
      }
    },
    "/crop/{harvestId}": {
      "post": {
        "operationId": "CropController_create",
        "parameters": [
          {
            "name": "harvestId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCropDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Crop"
        ]
      }
    },
    "/crop": {
      "get": {
        "operationId": "CropController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number to be returned (starts at 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (maximum 100)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaginationBodyDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Crop"
        ]
      }
    },
    "/crop/{id}": {
      "get": {
        "operationId": "CropController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Crop"
        ]
      }
    },
    "/crop/{id}/status/{status}": {
      "patch": {
        "operationId": "CropController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Crop"
        ]
      }
    },
    "/harvest/{farmId}": {
      "post": {
        "operationId": "HarvestController_create",
        "parameters": [
          {
            "name": "farmId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Harvest"
        ]
      }
    },
    "/harvest": {
      "get": {
        "operationId": "HarvestController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number to be returned (starts at 1)",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page (maximum 100)",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaginationBodyDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Harvest"
        ]
      }
    },
    "/harvest/{id}": {
      "get": {
        "operationId": "HarvestController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Harvest"
        ]
      },
      "patch": {
        "operationId": "HarvestController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateHarvestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Harvest"
        ]
      }
    },
    "/harvest/{id}/status/{status}": {
      "patch": {
        "operationId": "HarvestController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusHarvestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Harvest"
        ]
      }
    }
  },
  "info": {
    "title": "Brain Agriculture",
    "description": "Documentação do Brain Agriculture",
    "version": "1.0",
    "contact": {
      "name": "Igor Nascimento da Costa Silva",
      "url": "https://github.com/IgorNCS",
      "email": "igor.ncsilva@hotmail.com"
    }
  },
  "tags": [
    {
      "name": "brain-agriculture",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "Fulano"
          },
          "password": {
            "type": "string",
            "example": "senhasecreta123"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzUxMiIsInR5cCIgOiAiSldUIiwia2lkIjoiMjA5ZjhlNmMtZTVhNC00ZmFmLWE5ODItODkyMmRiNjVhN2Y3In0"
          },
          "refresh_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzUxMiIsInR5cCIgOiAiSldUIiwia2lkIjoiMjA5ZjhlNmMtZTVhNC00ZmFmLWE5ODItODkyMmRiNjVhN2Y3In0"
          },
          "expires_in": {
            "type": "number",
            "example": 1800
          }
        },
        "required": [
          "success",
          "access_token",
          "refresh_token",
          "expires_in"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzUxMiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICI2MjYxYjBmYS04ZjQ4LTQ0NzUtODhiNi0xMjIwNjg1YjZiZjUifQ.eyJleHAiOjE3NDM1NTUxNDYsImlhdCI6MTc0MzU1MzM0NiwianRpIjoiZDliMmNlYTktNWNjZi00YWQ2LTliNzktZWVhNmJhMTRmMjkxIiwiaXNzIjoiaHR0cDovL2xvY2FsaG9zdDo4MDgwL3JlYWxtcy9lbmN1cnRhZG9yIiwiYXVkIjoiaHR0cDovL2xvY2FsaG9zdDo4MDgwL3JlYWxtcy9lbmN1cnRhZG9yIiwic3ViIjoiOWQ1YmI4ZjctZTgxYS00ODdiLWFhMTYtNDE3ZDc1MTAxMzE1IiwidHlwIjoiUmVmcmVzaCIsImF6cCI6Im5lc3QtYXBpIiwic2lkIjoiM2ZiZGVjY2MtZDBiYy00ZDI3LTkyMjctMGU0NDMxY2ZlNzE3Iiwic2NvcGUiOiJiYXNpYyBwcm9maWxlIGFjciB3ZWItb3JpZ2lucyBlbWFpbCByb2xlcyJ9.U7MmoO0wapldVzTJquMajLSNWZwlKGdxAgc05-V3bDOvVgosCesGrCcanmZRUBCc_IT7JNBb7nStVWR3L6P3Hg"
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fulano de tal"
          },
          "email": {
            "type": "string",
            "example": "fulano@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "senhasecreta123"
          },
          "role": {
            "type": "string",
            "example": "costumer",
            "description": "Apenas ADMIN pode enviar"
          },
          "cpfCnpj": {
            "type": "number",
            "example": "12345678900"
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "cpfCnpj"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Fulano de tal"
          },
          "email": {
            "type": "string",
            "example": "fulano@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "senhasecreta123"
          },
          "role": {
            "type": "string",
            "example": "costumer",
            "description": "Apenas ADMIN pode enviar"
          },
          "cpfCnpj": {
            "type": "number",
            "example": "12345678900"
          }
        }
      },
      "CreateFarmDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Colheita feliz"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "example": "NY"
          },
          "totalArea": {
            "type": "number",
            "example": 100.5
          },
          "arableArea": {
            "type": "number",
            "example": 80
          },
          "vegetationArea": {
            "type": "number",
            "example": 20.5
          },
          "userId": {
            "type": "string",
            "example": "e7d3c6c4-3b6d-4e7a-9eae-64c93f9f7f4f"
          }
        },
        "required": [
          "name",
          "city",
          "state",
          "totalArea",
          "arableArea",
          "vegetationArea"
        ]
      },
      "PaginationBodyDTO": {
        "type": "object",
        "properties": {
          "userId": {
            "example": [
              "b7f0c5a0-0a1e-11ec-9621-0242ac130002"
            ],
            "description": "Client Id.",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "startDate": {
            "type": "string",
            "example": "2022-01-01",
            "pattern": "/^d{4}-d{2}-d{2}$/",
            "description": "Restricts the query to authorizations with Start Date greater than or equal to the informed. Format: yyyy-mm-dd"
          },
          "endDate": {
            "type": "string",
            "example": "2022-01-01",
            "pattern": "/^d{4}-d{2}-d{2}$/",
            "description": "Restricts the query to authorizations with End Date less than or equal to the informed. Format: yyyy-mm-dd"
          },
          "relations": {
            "type": "array",
            "example": [
              "user"
            ],
            "description": "Relations that will be returned.",
            "items": {
              "type": "string",
              "enum": [
                "user",
                "farm"
              ]
            }
          },
          "entityFilters": {
            "type": "object",
            "description": "Additional fields of the entity for filtering. Ex: { \"plantedArea\": 100 }",
            "additionalProperties": true
          }
        }
      },
      "UpdateFarmDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Colheita feliz"
          },
          "city": {
            "type": "string",
            "example": "New York"
          },
          "state": {
            "type": "string",
            "example": "NY"
          },
          "totalArea": {
            "type": "number",
            "example": 100.5
          },
          "arableArea": {
            "type": "number",
            "example": 80
          },
          "vegetationArea": {
            "type": "number",
            "example": 20.5
          }
        }
      },
      "CreateCropDto": {
        "type": "object",
        "properties": {
          "seed": {
            "type": "string",
            "example": "Milho"
          },
          "plantedArea": {
            "type": "number",
            "example": 20.5
          },
          "monthPlanted": {
            "type": "number",
            "example": 1
          },
          "yearPlanted": {
            "type": "number",
            "example": 2025
          }
        },
        "required": [
          "seed",
          "plantedArea",
          "monthPlanted",
          "yearPlanted"
        ]
      },
      "UpdateStatusHarvestDto": {
        "type": "object",
        "properties": {
          "monthHarvested": {
            "type": "number",
            "example": 5,
            "nullable": true,
            "description": "Month when the harvest was harvested."
          },
          "yearHarvested": {
            "type": "number",
            "example": 2025,
            "nullable": true,
            "description": "Year when the harvest was harvested."
          }
        }
      },
      "UpdateHarvestDto": {
        "type": "object",
        "properties": {
          "plantedArea": {
            "type": "number",
            "example": 20.5,
            "description": "Total area of the farm that was harvested."
          },
          "enabledArea": {
            "type": "number",
            "example": 80
          },
          "status": {
            "type": "string",
            "example": "preparing_land",
            "enum": [
              "preparing_land",
              "planted",
              "growing",
              "harvested",
              "canceled"
            ],
            "description": "Status of the harvest."
          },
          "monthPlanted": {
            "type": "number",
            "example": 1,
            "description": "Month when the harvest was planted."
          },
          "yearPlanted": {
            "type": "number",
            "example": 2025,
            "description": "Year when the harvest was planted."
          },
          "monthHarvested": {
            "type": "number",
            "example": 5,
            "nullable": true,
            "description": "Month when the harvest was harvested."
          },
          "yearHarvested": {
            "type": "number",
            "example": 2025,
            "nullable": true,
            "description": "Year when the harvest was harvested."
          }
        }
      }
    }
  }
}